requirements.txt
itsdangerous==1.1.0
fastapi_mail==0.3.6.1
python_jose==3.2.0
aiohttp==3.7.4.post0
pydantic==1.8.1
aiokafka==0.7.0
asyncpg==0.20.1
uvicorn==0.12.1
celery==4.4.7
email_validator==1.1.2
passlib==1.7.2
fastapi==0.65.1
jose==1.0.0
python-dotenv==0.17.1
secrets==1.0.2
